+++ storage_utils/swift.py
@@ -33,7 +33,6 @@ import cStringIO
 from functools import wraps
 import json
 import logging
-import operator
 import os
 import posixpath
 import tempfile
@@ -85,9 +84,8 @@ If not set, the ``OS_TEMP_URL_KEY environment variable will be used.
 # Make the default segment size for static large objects be 1GB
 DEFAULT_SEGMENT_SIZE = 1024 * 1024 * 1024
 
-# When a swift method takes a condition, these variables are used to
-# configure retry logic. These variables can also be passed
-# to the methods themselves
+# These variables are used to configure retry logic for swift.
+# These variables can also be passed to the methods themselves
 initial_retry_sleep = 1
 """The time to sleep before the first retry"""
 
@@ -178,6 +176,16 @@ class UnauthorizedError(SwiftError):
     pass
 
 
+class ConflictError(SwiftError):
+    """Thrown when a 409 response is returned from swift
+
+    This error is thrown when deleting a container and
+    some object storage nodes report that the container
+    has objects while others don't.
+    """
+    pass
+
+
 class ConfigurationError(SwiftError):
     """Thrown when swift is not configured properly.
 
@@ -192,88 +200,28 @@ class ConditionNotMetError(SwiftError):
     pass
 
 
-class _Condition(object):
-    """A conditional expression that can be applied to some swift methods
+def _validate_condition(condition):
+    """Verifies condition is a function that takes one argument"""
+    if condition is None:
+        return
+    if not (hasattr(condition, '__call__') and hasattr(condition, '__code__')):
+        raise ValueError('condition must be callable')
+    if condition.__code__.co_argcount != 1:
+        raise ValueError('condition must take exactly one argument')
 
-    Condition objects take an operator and a right operand. Left operands
-    can be checked against the condition. If the condition is not met,
-    ConditionNotMetError exceptions are thrown.
 
-    The operators that are supported are ``==``, ``!=``, ``<``, ``>``, ``<=``,
-    and ``>=``.
+def _check_condition(condition, results):
+    """Checks the results against the condition.
 
+    Raises:
+        ConditionNotMetError: If the condition returns False
     """
-    operators = {
-        '==': operator.eq,
-        '!=': operator.ne,
-        '<': operator.lt,
-        '>': operator.gt,
-        '<=': operator.le,
-        '>=': operator.ge
-    }
+    if condition is None:
+        return
 
-    def __init__(self, operator, right_operand):
-        if operator not in self.operators:
-            raise ValueError('invalid operator: "%s"' % operator)
-        self.operator = operator
-        self.right_operand = right_operand
-
-    def assert_is_met_by(self, left_operand, left_operand_name):
-        """Asserts that a condition is met by a left operand
-
-        Args:
-            left_operand: The left operand checked against the condition
-            left_operand_name: The name of the left operand. Used when
-                creating an error message
-
-        Raises:
-            ConditionNotMetError: When the condition is not met
-        """
-        if not self.operators[self.operator](left_operand, self.right_operand):
-            raise ConditionNotMetError(
-                'condition not met: %s is not %s' % (left_operand_name, self)
-            )
-
-    def __str__(self):
-        return '%s %s' % (self.operator, self.right_operand)
-
-    def __repr__(self):
-        return '%s("%s", %s)' % (type(self).__name__,
-                                 self.operator,
-                                 self.right_operand)
-
-
-def make_condition(operator, right_operand):
-    """Creates a condition that can be applied to various swift methods.
-
-    Args:
-        operator (str): The operator for the condition, can be
-            one of ``==``, ``!=``, ``<``, ``>``, ``<=``,
-            and ``>=``.
-        right_operand: The operand on the right side of the
-            condition.
-
-    Examples:
-        To make a condition and use it in `SwiftPath.list`::
-
-            from storage_utils.swift import make_condition
-            from storage_utils.swift import SwiftPath
-            p = SwiftPath('swift://tenant/container/resource')
-            # Ensure that the amount of listed objects are greater than 6
-            cond = make_condition('>', 6)
-            objs = p.list(num_objs_cond=cond)
-
-        An illustration of what it looks like when a condition doesn't pass::
-
-            cond = make_condition('>', 100)
-            # ConditionNotMetError is thrown when the condition is not met
-            objs = p.list(num_objs_cond=cond)
-            Traceback (most recent call last):
-            ...
-            storage_utils.swift.ConditionNotMetError: condition not met: num
-            listed objects is not > 100
-    """
-    return _Condition(operator, right_operand)
+    condition_met = condition(results)
+    if not condition_met:
+        raise ConditionNotMetError('condition not met')
 
 
 def _swift_retry(exceptions=None):
@@ -319,6 +267,8 @@ def _propagate_swift_exceptions(func):
                 raise UnauthorizedError(str(e), e)
             elif http_status == 404:
                 raise NotFoundError(str(e), e)
+            elif http_status == 409:
+                raise ConflictError(str(e), e)
             elif http_status == 503:
                 logger.error('unavailable error in swift operation - %s', str(e))
                 raise UnavailableError(str(e), e)
@@ -779,14 +729,14 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
     def list(self,
              starts_with=None,
              limit=None,
-             num_objs_cond=None,
+             condition=None,
              # intentionally not documented
              list_as_dir=False):
         """List contents using the resource of the path as a prefix.
 
         This method retries `num_retries` times if swift is unavailable
         or if the number of objects returned does not match the
-        ``num_objs_cond`` condition. View
+        ``condition`` condition. View
         `module-level documentation <swiftretry>` for more
         information about configuring retry logic at the module or method
         level.
@@ -796,9 +746,8 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                 be appended to the resource of the swift path. Note that the
                 current resource path is treated as a directory
             limit (int): Limit the amount of results returned
-            num_objs_cond (SwiftCondition): The method will only return
-                results when the number of objects returned meets this
-                condition.
+            condition (SwiftCondition): The method will only return
+                when the results matches the condition.
 
         Returns:
             List[SwiftPath]: Every path in the listing.
@@ -806,12 +755,13 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         Raises:
             SwiftError: A swift client error occurred.
             ConditionNotMetError: Results were returned, but they did not
-                meet the num_objs_cond condition.
+                meet the condition.
         """
         connection = self._get_swift_connection()
         tenant = self.tenant
         prefix = self.resource
         full_listing = limit is None
+        _validate_condition(condition)
 
         # When starts_with is provided, treat the resource as a
         # directory that has the starts_with parameter after it. This allows
@@ -848,9 +798,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             for r in results[1]
         })
 
-        if num_objs_cond:
-            num_objs_cond.assert_is_met_by(len(paths), 'num listed objects')
-
+        _check_condition(condition, paths)
         return paths
 
     def listdir(self):
@@ -862,7 +810,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         return self.list(list_as_dir=True)
 
     @_swift_retry(exceptions=(ConditionNotMetError, UnavailableError))
-    def glob(self, pattern, num_objs_cond=None):
+    def glob(self, pattern, condition=None):
         """Globs all objects in the path with the pattern.
 
         This glob is only compatible with patterns that end in * because of
@@ -874,7 +822,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         method will perform a swift query with a prefix of mydir/pattern.
 
         This method retries `num_retries` times if swift is unavailable or if
-        the number of globbed patterns does not match the ``num_objs_cond``
+        the number of globbed patterns does not match the ``condition``
         condition. View `module-level documentation <swiftretry>`
         for more information about configuring retry logic at the module or
         method level.
@@ -882,9 +830,8 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         Args:
             pattern (str): The pattern to match. The pattern can only have
                 up to one '*' at the end.
-            num_objs_cond (SwiftCondition): The method will only return
-                results when the number of objects returned meets this
-                condition.
+            condition (SwiftCondition): The method will only return
+                when the number of results matches the condition.
 
         Returns:
             List[SwiftPath]: Every matching path.
@@ -892,18 +839,17 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         Raises:
             SwiftError: A swift client error occurred.
             ConditionNotMetError: Results were returned, but they did not
-                meet the num_objs_cond condition.
+                meet the condition.
         """
         if pattern.count('*') > 1:
             raise ValueError('multiple pattern globs not supported')
         if '*' in pattern and not pattern.endswith('*'):
             raise ValueError('only prefix queries are supported')
+        _validate_condition(condition)
 
         paths = self.list(starts_with=pattern.replace('*', ''), num_retries=0)
 
-        if num_objs_cond:
-            num_objs_cond.assert_is_met_by(len(paths), 'num globbed objects')
-
+        _check_condition(condition, paths)
         return paths
 
     @_swift_retry(exceptions=UnavailableError)
@@ -1062,11 +1008,11 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                  dest,
                  object_threads=10,
                  container_threads=10,
-                 num_objs_cond=None):
+                 condition=None):
         """Downloads a directory to a destination.
 
         This method retries `num_retries` times if swift is unavailable or if
-        the number of downloaded objects does not match the ``num_objs_cond``
+        the number of downloaded objects does not match the ``condition``
         condition. View `module-level documentation <storage_utils.swift>`
         for more information about configuring retry logic at the module or
         method level.
@@ -1081,20 +1027,22 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                 objects.
             container_threads (int): The amount of threads to use for
                 downloading containers.
-            num_objs_cond (SwiftCondition): The method will only return
-                results when the number of objects downloaded meets this
-                condition. Partially downloaded results will not be deleted.
+            condition (SwiftCondition): The method will only return
+                when the results matches the condition. In the event of the
+                condition never matching after retries, partially downloaded
+                results will not be deleted.
 
         Raises:
             SwiftError: A swift client error occurred.
             ConditionNotMetError: Results were returned, but they did not
-                meet the ``num_objs_cond`` condition.
+                meet the condition.
 
         Returns:
-            dict: A mapping of the destination to the download path
+            List[SwiftPath]: Every downloaded path.
         """
         if not self.container:
             raise ValueError('cannot call download on tenant with no container')
+        _validate_condition(condition)
 
         service = self._get_swift_service(object_dd_threads=object_threads,
                                           container_threads=container_threads)
@@ -1107,11 +1055,8 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                                            self.container,
                                            options=download_options)
 
-        if num_objs_cond:
-            num_objs_cond.assert_is_met_by(len(results),
-                                           'num downloaded objects')
-
-        return {self: dest}
+        _check_condition(condition, results)
+        return results
 
     @_swift_retry(exceptions=UnavailableError)
     def upload(self,
@@ -1224,7 +1169,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                                         self.container,
                                         [self.resource])
 
-    @_swift_retry(exceptions=UnavailableError)
+    @_swift_retry(exceptions=(UnavailableError, ConflictError))
     def rmtree(self):
         """Removes a resource and all of its contents.